PSYNEU-81 feat: Add withClipPath higher order function
